+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 Changelog
 ---------
 
+0.5.6 (2015-02-26)
++++++++++++++++++++
+
+* fix "苹果" pinyin error. `#11`__
+* 精简 phrases_dict
+* fix 重复 import jieba 的问题
+* 更新文档
+
+__ https://github.com/mozillazg/python-pinyin/issues/11
+
+
 0.5.5 (2015-01-27)
 +++++++++++++++++++